reorder-agent-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
prometheus-client==0.19.0
apscheduler==3.10.4
httpx==0.25.2
python-json-logger==2.0.7
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.13.1
tenacity==8.2.3
pytest==7.4.3
